MDEV-27936 hardware lock ellision on ppc64{,le} failing to compile
There is only a very small range of gcc compiler versions
that allow the built_{htm} functions to be defined without -mhtm
being specified as a global CFLAG. Without this cflags the
__built_tbegin and other htm instructions simply aren't available.

The include header htmxlintrin.h uses these inlines. As such
the file that includes it need to be compiled with -mhtm. Other
tricks like #pragma GCC target("cpu=power8","htm") around the
include block had no effect on making the builtins avialable.

As such we are left with the option of putting the include header
in one location, srw_lock, and enabling it there.

As a partial mitigation, extented to GNU/clang compilers,
transaction functions gain the attribute "hot".

In general the use of htm is around the optimistic
transaction ability of the function. The key part of using the
hot attribute is to place these functions together so that
a maximization of icache, tlb and OS paging can ensure that
these can be ready to execute by any thread/cpu with the
minimium amount of overhead. The increased speed on
the transaction to not be conflicted and rollback on tend.

trx_has_lock_x cannot be cold and hot so settled on hot.

POWER is particulary affected here because the xbegin/xend
functions are not inline.