MDEV-27923 Deprecate spider_use_handler
Deprecate the server variable spider_use_handler and the corresponding
table parameters "uhd" and "use_handler".

Passing a Handler statement to data nodes, without converting it to
SQL sometimes, might improve the performance, while this introduces
some complication to the implementation.

In the first place, only a few people use Handler statements and the
performance gain seems not to be very significant. Further, setting
spider_use_handler > 0 disables the GROUP BY handler. So, we decided
to deprecate the variable.